Rendering Improvements
- Improved FBO resize using the attachments
- Using the same depth buffer between gbuffer and renderbuffer to avoid blitting or passing. Works great!
- Added support for object mouse picking using the gbuffer
- Enhanced OpenGLAPI with more methods

+++ NibbleCore/Core/Entity.cs
@@ -13,7 +13,7 @@ namespace NbCore
         public ulong NameHash;
         public string Path = ""; //Entity path (if any)
         public EntityType Type;
-        public ulong ID = 0xFFFFFFFF;
+        public uint ID = 0xFFFFFFFF;
         public long testID = 0; //TODO: remove that when we're happy with memory disposal
         public bool Initialized = false;
         public static long test_counter = 1;
@@ -33,7 +33,7 @@ namespace NbCore
             testID = test_counter++;
         }
 
-        public void Init(ulong id)
+        public void Init(uint id)
         {
             if (Initialized)
                 return;
